+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ------
 * [#1298](https://github.com/Shopify/shopify-cli/pull/1298): Fix error in `theme serve` command
 * [#1301](https://github.com/Shopify/shopify-cli/pull/1301): Add `theme init` command
+* [#1295](https://github.com/Shopify/shopify-cli/pull/1295): Ignore files at the root of a theme app extension project
 
 Version 2.0.0
 -------------